refactor: split reconcile runtime helpers
